Quite a number of changes all put in as one commit. Sorry.
Added a quick hack to avoid running includable tests on all non-erroneous subjects.

The AST_hir_if, we decided, was to be changed to a branch, which is conditional, and a goto, which isnt. Having done this, and written a few transforms that use it, I think that the old way might be better, since it allows the goto code to only be handled once. That said, this is done, so it might be better to leave as is.

Added a new goto_uppering plugin, which undoes the changes from the goto_lowering pass. It cant really undo the changes, but it simulates it with a while statement, a state variable, a switch statement, and a large number of continue statements. In theory it should work, but it hasnt been tried for two reasons:
	- Firstly, plugins are run before lowering, so there are no gotos to be transformed. This is part of a larger problem that we dont run command line options in the order given, but rather an arbitrary order which made sense at the time (it mostly still does). Since gengetopt doesnt give us an option to get the options in order, im not sure what to do here.
	- Secondly, the code relies on the function splitting pass, which I think Edsko is writing tomorrow, so there no point hacking this to work otherwise.
	- As such, note that the code is not tested (it compiles), and I'm fairly sure its not linear (search for l1->value->value). I havent seen the linearity test in a long time, so ill check if it fails here before fixing it.

Updated the Interpret shredded test to use the new goto_uppering plugin.