Slightly different setup in the code generator. Previously, when we assigned a zval* to a local variable, we did not count the reference to the zval* as a reference. That means that when we created an integer, we set its refcount to 0, and when we did something like
  zval* rhs = index_ht(..)

to get the zval from a hashtable, we used whatever refcount the variable
already had in the hashtable. Then at the end of an assignment, we checked if
the RHS had refcount 0, and if so, deleted it.

This structure worked but was arguably slightly counter-intuitive and moreover
was difficult to work with when implementing function calls. Instead, we now
treat the local references as real references. Thus, when we create an integer,
we leave its refcount at 1, and when we index a hashtable, we increment the
refcount. At the end of an assignment we can simply call zval_ptr_dtor on the
RHS, which will decrement the refcount of the RHS and garbage collect it when
it drops to zero.

This should have solved the memory leak for returning by reference, but
unfortunately I had to add a hack to work around a problem in the Zend engine
(which resets the refcount and is_ref field of the return value of a function,
which obviously causes problems). It should now work in most (but not all)
cases-- but will result in memory leaks in some cases, and in actually invalid
behaviour (segfaults?) in others. Not sure how to solve this one; awaiting
reply on the PHP internals mailing list (don't hold your breath).