Using references everywhere is subtly wrong, so places where references are required are noted, and non-references are used elsewhere. An example to show the subtlty is
	$x[0][0];
	var_dump ($x);

$x should be NULL, but if we use references, it will actually be an array with a NULL element.

We now add references in the following cases:
	$x =& $y; // reference for RHS
	function &x () { return $x; } // when we add the temporary for $x, use a reference
	list ($x) = $y; // use reference if $y is not a literal

We also move the annotation to after the early shredding, since that creates some variables which need to be annotated.

Finally, fixed the purity_test, which hasnt worked for ages. It still isnt regularly tested, but there is now a subject (impure.php) which it can be manually tested on.