Variables accessed as arrays would fail if they weren't arrays. The correct thing to do is set the lhs to NULL (except in the case of strings, which isnt yet handled).
This brings up the problem of deleting and adding to the hashtable instead of updating. It would break this code:

$x = array ("a" => 1, "b" = 2, "a" = 3);

since it would move "a" to the end of the array, and it would iterate in the wrong order.

I believe it also causes a few infinite loops in foreaches. I dont understand fully why this is, but I expect the internal iterator is being messed with when we delete, but not when we update.