The name no_create was inaccurate. We were really creating zvals which weren't added to the hashtables. Renamed some functions to indicate this.
When the _dont_add functions were called, their refcounts got set a little high. We now decrement them immediately (decrementing them outside the function is impossible, since we've lost the context of whether they're found in the hashtable or not).

Indexing and updating the hashtable was a little bit off. We must handle bools as longs, use "" for NULLs, and convert doubles to ints, before inserting or fetching. Also, added indenting for some generated code, to be able to read it better.