A lot of the necessary process_includes work is done here.
A number of extra test cases, especially testing warnings, are added.

The shredder is altered to not shred the parameters to include (), require (), etc. Since we dont have any dataflow, if we separate the filenames from the method invocation, we wont be able to find out the file to include.

When doing the late inclusion, we need to run the passes on the included script up until the same point. phc.cpp now supplies the pass_manager and name of the pass to run up until.