This implements the new strategy. Instead of generating bits of code, and hoping it all works together, with copious scope marks thrown in for good measure, its easier to understand and debug to use a function for each logical step, and to only pass around the necessary state. The only state thats needed for a variable is a zval* and a boolean to say whether we have created a variable or not (as we'll need to destruct it if we have).
Fetching of variables is split into read_var and write_var, which will determine which form of reading/writing the variable is required (array, scalar, array_push or variable_variable). Some functions with old functionality are still around to preserve compatibility of things we havent looked at yet.

So far, simple variables, and actual parameters are supported. It shouldnt be too much work to get back on track with the other functions.

There is an unusual change involving EG(uninitialized_zvap_ptr). Instead of allocating memory with MAKE_STD_ZVAL, we assign the pointer to this function. The run-time recognizes this (apparently, I've seen it in actual parameters, but it works elsewhere). The only real clue to this is that the refcounts after a function call are sometimes 1, but only if "zvp == EG(uninitialized_zvap_ptr)".