Add read_array_index and write_array_index. they work, but dont have correct refcounts.
A number of bugs were occuring because I had

if (...)
	...;
else
	ALLOC_INIT_ZVAL (zvp);

ALLOC_INIT_ZVAL is actually two statements, so needs to be wrapped in curlies.

Am trying to set it up so we can fail on gcc warnings. Marked the end_of_function label as unused in order to do this.