So far I havent been certain that my handling of EG(uninitialized_zval_ptr) has been correct. In particular, I feel some tests may overwrite this, and still pass.
This makes EG(uninitialized_zval_ptr) point into read-only memory, so updating it will lead to a segfault. This breaks about half the tests, bringing the bugs out into the open. It is possible that some bugs may occur due to the fact that EG(uninitialized_zval_ptr) no longer points to EG(uninitialized_zval), in which case I'll have to remove this later.

Note that the 2 is intentional, as PHP increases the refcount to ensure that it is always separated.