The main work of this commit was to set up a constant pool, and preallocate values into it.
We allocate 1 value per Token_int, Token_real, or Token_bool of that value (so the pool is shared for all int(0)s, for example). Token_null just uses EG (uninitialized_zval_ptr). Token_string is not done yet.

Up to now, about 27% of the time was spent in the allocator. Now only 20% is. This correlates with the benchmark results of about a 10% speedup with this.

Interestingly, there was almost no speedup when this was only applied for ints. Since there are almost no bools or NULLs in the program, I can only assume the reduction was partly due to the reduction of zend_eval_string, as well as the allocation problem.


Unfortunately, I had to mess around with the stringstreams a little. I added a prologue so that we could add the phc_const_pool_* declarations there, since the need to be provided before use. It might be cleaner to have a separate pass, but it was nearly done this way.

We also replace "cout << " with "code << ", which removes a bit of a hack that we had before.


This makes all the refcounts go wrong, unfortunately, making the test a bit poor. I will add a flag for this so that refcounts can still be useful.