Modified the grammar so that we have an explicit construct for "opeqs". We used to represent
  $x += $y;

as

  $x = $x + $y;

in the grammar, but that is not valid in the general case:

  $x[f()] += 1;

cannot be represented using a binary operator, but that would cause f() to be
called twice. This also fixes a bug in the shredder where opeqs were only
properly dealt with as top-level expressions (see opeq2.php for a
counterexample that broke).