After each pass, the consistency check runs to ensure that we havent created an illegal PPH construct. If such a structure occurs, then we get an internal error. However, first, we must run the invalid check, so that if those errors are from the source file, that we report those immediately, instead of giving an internal error, or segfaulting later. So invalid check needs to be the first pass.
However, we want the list_passes pass to go first, so that we can list the passes even if there are errors. Before this commit, the consistency check after list_passes meant that use errors became internal errors.

Instead, we not have a 'check' flag in the Pass_manager, and it will not run the consistency check unless this is set. We also remove the consistency check, and reuse the internal check instead (which is sort of what was going on anyway).