Replace expressions with variable names in variable array_indices and replace variables in actual_parameters with the same fields as variables have.
Move debug () functions to debug.{cpp,h}, away from AST_unparser and HIR_unparser.

The grammar changes above aren't reflected in the HIR_unparser, as it currently has no test.

I tried to fix the Process_includes transform so that if a file wasnt included, it would transform the include function to use variable_names in the parameter. However, the pass was a mess, so it got seriously refactored as well. (This currently has no proper test, which is a problem.)

In Generate_C, added a phc_unsupported function, to differentiate between asserts and things which need to be fixed.