Replace the xml_roundtrip with one which reads from a phc instance and writes the the next one.
We construct a long series of piped command lines, such as
	phc --read-xml=pass1 --xdump=pass2 | phc -- read-xml=pass2 --xdump=pass3 ...

The final result should be the same as if it were run straight through phc without being dumped to XML. 

In actual fact, we fail in a number of places, which should be fixed. The primary lpace we fail is in names of temporary variables. We should attempt to fix that in the future.