This provides everything needed to put a simple hello-world PHP program into the syntax database.
There is a huge amount of boiler-plate here, but the types aren't really required at all, so hopefully this would be do-able without them. That would allow the XML parser to just create the tags and types in the databases as they go, and not worry about defining the types in the front-end.