Fix up the reduce script to be more useful and correct.
Firstly, reduce was designed to test the code generation. However, it's equally useful to debug problems earlier in the compilation. Add a -i flag, signifying interpretation, to dump the output at a certain pass and interpret the output.

If there's a bug in the upperer, we can use the -U flag will instruct reduce to never use it.

Add a check at the start that there even is a potential bug we can find. Otherwise we'll be chasing ghosts.

The program used read-ast-xml, which is defunct. Replaced with -read-xml=ast.

Added more debugging info at -vv.

Close pipes at program end.