Fix the Dead_code and copy_propagation passes to work at the global scope, as well as in functions. This is almost a complete rewrite of the use-def portion.
The use_def_counter now uses a stack to avoid the nested-functions problem. Instead of the nasty hack of being passed a string to int map, it now annotates each variable with
in a function with its uses and defs. That is really the point of the attrs maps anyway. These are cleared by the use of the pass, in Clear_use_defs.

We also remove an assertion, where I assumed that left-hand sides of replaceable statements are unique. This doesn't work in the case of If statements with the same lhs in bo
th branches. This occurs when trinary elements are lowered. Fortunately, this doesn't affect correctness.