Add constant folding. We check for bin_ops and unary_ops with only literals as their expressions, and unparse the expression into a string, which is passed to the PHP run-time. The value is then extracted from the zval.
The bulk of the work is in embed.cpp. We wrap zend_eval_string in zend_try and zend_catch, and use handle_php_error_cb as an error handler. Parse and other errors are caught by the handler, which then prints a warning, and throws zend_bailout(), taking us directly to zend_catch(). In this case, we return the original expr, which is why the signature uses Expr as its return type, not Literal.

If the PHP embed SAPI is missing, we just return the passed Expr. This hasnt been tested, or even compiled.

misc/constant_fold_test_generator.php was used to create all of the tests. Tests which failed or warned were moved to separate files in warnings/ and errors/. I also manually wrote a few compositions of tests, which didnt cause any errors.

Constant_folding re-adds checks for double '-' unary_ops, which I had removed from the AST cleanup.