When XML is created from one instance of phc, then and read into another, the suffix for fresh() variables and labels is reset to zero, meaning any new fresh variables will use the same name, which is a bad idea. This searches through the list of suffices, and sets the counter to the highest found, +1.
This isnt really the optimum solution to the problem. A better idea would be to get a list of all variables in the program, and make sure we don't overwrite them. However, this is done, and good enough for now. However, I added a prefix to those variables: __phc__. This might be annoying in practice, but we'll see later.

Note that this doesnt completely fix the XML test, as Foreign nodes are not properly printed. This is an exercise for another day.