Do most of the work to convert to using sum-types rather than predicates. In most cases, this makes things simpler and more elegant, and avoids relying on the mir() database after CFG construction.
This doesnt copy source_rep and attr predicates. They will be done shortly.

Most changes are mechanical changes from, for example:

	get_method_name (METHOD, NAME) :- 
	mir()->method (METHOD, SIG, _), 
	mir()->signature (SIG, _, _, NAME_ID, _), 
	mir()->mETHOD_NAME (NAME_ID, NAME).

to 

	get_method_name (METHOD, METHOD_NAME) :- 
	  METHOD = method{_, signature{_, _, _, mETHOD_NAME{_, METHOD_NAME}, _}, _}.


The CLPA_unparser has gone through large changes. In particular, we no longer need to do a complicated stack traversal, and a simple visitor with minor hacks is all that required. Most of the work is done is pre_/post_node. There are some minor hacks, as we'd prefer to have visit_optional as pre/post_optional, and the same for visit_type. However, I looked at doing this in maketea, and it would have been difficult.