This is the first version where Dead-code elimination works (at least for simple cases).
We switch to a listS for the BGL graph. Using a vecS (vector store) instead of a list means that removing a vertex breaks the reverse mapping of bb->vertex.

This implements replace_bb.

init_df and should_iterate really belong to the analysis rather than the BB. I think it will be necessary to have a "Solution" class that knows when it has been updated, (this is what a basic_block was intended to be, but now it looks like it will be a solution container).