Testing individual features was nasty. It relied on a 6 field comma separated value that I found it very difficult to get right, and had to look up every time. This replaces it with dejaGNU style annotations. Comments have braces with key-value pairs, separated by colons. Example annotations:
	/* { phc-error: Cannot use [] in strings } */
	// { phc-option: --include }

Importantly, we separate this from the basic_parse_test. Now, the basic parse tests catch anything that has been missed by annotations, such as un-annotated errors. (But it wont check options, perhaps it should?).

This makes it easier to do precise checks of certain features, without the huge overhead of the brute-force checks we currently have.