Adds Param_is_ref. This allows a function call like:
	f($x[$i]);

to be replaced with:

	if (param_is_ref (f, 0))
		$T =& $x[$i];
	else
		$T = $x[$i];
	f ($T);

This means that method_invocation only has simple parameters. It also allows an optimization pass to replace fold param_is_ref in the (hopefully large) number of cases where we know the answer ahead-of-time.

Isset and Unset are now special in the MIR, as they require non-shredded parameters. They were matched seperately in Generate_C anyway, so this doesn't change matters all that much.

Not all tests pass, but enough pass to commit this, and clean up the few fails later. The cb_mir fail is unrelated, its just exposed with this patch.

Param_is_ref_uppering works, but it shouldn't, as we dont use ReflectionClass, or take targets into account. I'll fix this as soon as I find a test which fails.