This fixes two related problems. Firstly, the logs are getting massive, and they are rarely used. This gzips them when they are created.
Secondly, the testing was a little bit off. We should be doing a long test, and we should be doing it in 'install' mode, since that is what really matters (in-tree tests get run all the time anyway).

This tidies the details page a little bit, since we now only have one column. I reduced the border too, to squeeze another row onto the page.

For my next trick, I'll redo the (recent) old tests.