Compile dynamic methods, classes and interfaces. This moves the dynamic declaration to the top-level, with a different name, replacing it with a call to {Class,Method,Interface}_alias. When the alias is executed, it sets the symbol table entry for the aliased function. It probably needs some error checking, but it handles all the cases we have in the tests.
This adds alias uppering and lowering, but it still allows classes etc to be statements in the MIR. That could be removed, but its complicated at first glance.

Classes cant be nested (stopped statically at decalration time). We allow classes to be nested, but not by design. We should add a check in Invalid_check. I added a test for this.