The CFG updating code is getting complex, and has bugs. It needs to be refactored to remove the redundancy that led to those bugs, and to really understand whats going on. First, I should commit what I have:
This fixes the dominance_frontiers. Lots were being missed. I'm not sure what the bug was, as I replaced the frontier creation and propagation with the Cooper/Torczon method, which is significantly shorter and easier. (This exposed the other bugs).

Support --cfg-dump for passes ssa, cfg and outssa.

Only dump_graphviz(NULL) if enable_debugging is set.

Check phi nodes in the consistency check.

Add dominance::dump().

Half-baked changes to debug the problems in updating the CFG.