Fix, in nearly all cases, updating the phi nodes when edges change.
So far, this method of keeping the phi node edges updated has been a disaster, being nearly as difficult as keeping the graph in sync, which we had nicely avoided by using boost.

I'm going to experiment with moving the phi arguments to edge properties. They'll then go away wehn edges are removed, and can be moved to the new edges from the old ones if necessary.