Add function signatures for PHP functions. I've only used them in one place so far: adding aliases for actual function parameters only if the call is by reference. Though I can see a number of places this has reduced the number of aliases, the performance results remain the same.
This introduces an optimization Oracle, which sits between PHP and the optimizers. It interfaces with PHP to get function signatures, but can also be given functions signatures by the optimizers.

An MIR::Signature now has an extra field: pass_rest_by_ref, indicating that parameters that do not have a formal parameter are passed by reference. Even though we cant create functions with the property (or I dont think we can), internal functions can use this, and so we need to model it.

We don't need to fix_solo_phi_args since we drop phi args. We can model dataflow properties through phis just as well as if they were put into real statements, but the real statements wouldnt be dropped at the end of the function.