Add an optimization to speed up loops controlled by boolean variables. All our for-loops appear like this in the MIR, due to weird lowering during conversion to the HIR. It is not possible (AFAICT) to fix this in the HIR, because lowering to a more efficient form will always fail at an edge case (typically involving a continue statement).
This pass identifies if statements with two predecessors, where the block variable is set to true in one predecessor, and set to false in one successor. In this case, it splits the control block. Each control block will only have one possible value, so SCCP and DCE will remove the boolean assignments and the control blocks.

This reduces the time spent in the benchmark by about 4%.

A number or organizational changes were required:
	Optimization_pass now has a require_ssa field. Without it, SSA wont be built (the Remove_loop_booleans pass doesnt need it).
	List now has a has() method, which returns whether an element is in the list.
	Basic_blocks now have a clone method, though this is only implemented for Branches, for now.

Some minor bug fixes are included:
	debugging information before cfg_dumnp was called didnt have any block numbers (they all used 0).
	has_def didnt work for non-SSA form.