Merge changes to Generate_C from dataflow to trunk.
Principally, this allows rvalues on the RHS of expressions. This is allowed on the dataflow branch due to constant propagation, but not on trunk. However, I need to sychronize the two, so that the refactoring of Generate_C is not too difficult to merge back to dataflow, and takes the changes into account.