Completely remove the LIR, and fix codegen to work without it.
The pass formerly known as Generate_C is now Generate_C_pass, and it calls Generate_C.

Generate_C no longer spits of LIR::Pieces. Instead, each pattern gets a 'buf', and compile_c combines the strings from the 'buf's into a larger body. Technically, each pattern uses a buf just as they did the 'code' buffer, but I've renamed it to not overload it with the old 'code' buffer. With the exception of adding a few buffer parameters, Generate_C changes very little due to this.