Add a new feature. When calling a macro or callback 	LHS#is_initialized is a clone of LHS with the is_initialized attribute set. These can also be chained: 	LHS#is_uninitialized#is_ref.
This should reduce a lot of the code duplication thats been going on (and a lot of the duplication I was going to add).


This also reworks the MICG_parser to use a parse_tree (as the Spirit AST has the not-very-useful property that nodes which only have one child node are omitted).

This also adds MICG to the XML_unparser, as this greatly aids debugging.