This separates the testing from the benchmarks in the nightly test, which will allow me redo the benchmarks without spending hours per commit on the tests.
This also completely redoes the DB layout, though its not immediately obvious from the code. Meta-results, such as time taken etc, are put into the components table, and there is a result per-component (where the components are compile, test and benchmark).

Results are kept (where they were) in the benchmarks/tests table, and duplicated information is removed from the complete table. I've made a reasonable effort to go through and tidy up poor function- and variable-names, and comments.

This also updates the web front-end to use the data, though thats not 100% finished.