Fix issue 102: phc won't build if HAVE_EMBED is not set. A number of methods had no defintion, or the definition had no signature.
When that bug was fixed, the cb_sua test failed for most files. PHP::convert_token is called to give an INT or a REAL a value (the parser assigns them 0 initially). Using the non-HAVE_EMBED versions, it did not add this value (except for overflowing integers). This adds it.

Tested for './configure' and './configure --without-php'.