This divides a points-to graph into 3 types of edge: 	- a field-to edge (storage -> index) 	- a reference edge (index -> index) 	- a points-to edge (index -> storage)
This makes the certainty field of an Alias_pair implicit in some cases. More importantly, it causes the Points_to API to be changed to be a lot more explicit, which is a good step. The lower-level API of Points-to is likely to vanish as a result.

The next step is to change the clients of Points-to and Aliasing to reflect the changes to the structure. This should make things simpler, and more correct.

Most of these changes aren't yet implemented, but it compiles now.
There is also a change in capitalization of Certainty (now uppercase C).