Fix merging of values in Pair_map, which makes references work properly.
I had been simulateously been trying to keep references in transitive-closure, and to avoid unification of points-to sets. However, these aren't mutually exclusive, I had been acting liek they were. Fortunately it wasn't implemented, but when I just implemented it now, I realized the problem. Added documentation to that effect.

Is assign reference, $x =& $y; a reference is created between $x and $y, but also between all of the references of $y and $x. References are never copied from $x to $y, even if the assignment to $x is not a killing definition.

Add checks so that Certaintys are valid.