There was contention over whether to remove a field edge. In the case of killing references, we should not have left field edges. In the case of assignments to varaibles with references, it is incorrect to remove it. Other assignments don't care.
Clearly it is the responsibility of kill_reference to remove the edge. But this relies on kill_reference being called after kill_value (which was not the case for unset and assign_by_ref).
Rather than comment this dependency, it is better to combine the two (kill_reference was always called with kill_value). As such, kill_reference becomes a boolean parameter of kill_value, defaulting to false.