I had been killing definitions before constructing the values to be assigned to them. This caused identity operations to fail. We now copy a RHS's values into a fake index node (FAKE::fake), from which the values are copied into a LHS (potentially after killing the old value). This could do with a tidy-up, however.
This also fixes a larger problem, which is that a single name was chosen for any memory allocated in a statement. This is irksome when copying an array, as the entire array will be collapsed into one name. Instead, we use one name per create_empty_storage call, using a counter. It seems that these are copied in the same order every time, and that the analysis terminates as a result, but I am not certain.