Added new visitor class: Stat_collector, for gathering statistics.
Updated Makefile.am/Makefile.in to reflect this.

Added a Stat_collector object to Whole_program along with a function, collect_stats (), to make use of it.

Modified increment_stat () in stats so that it returns an int, the value before the stat was incremented (i.e. it'll return 0 the first time the counter is incremented, so we can tell, for example, if a new type has been assigned to a variable.).
Modified CTS () so that it works in an if statement, i.e. it's no longer wrapped in a do{}while(0);
(I might revert the two above changes to the way they were and write a seperate function/macro for returning the value of stat counters, but for now that works)

Stats weren't being dumped with --stats, so phc.cpp now dumps stats automatically after every time it runs

Due to the above, dump_stats () now writes to cerr rather than cout (dumped CFGs were getting corrupted when cout was used)